MVC식으로 정리 및 오류 수정
